pipeline/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
pymupdf==1.24.5
openai==1.35.13
google-generativeai==0.7.2
python-dotenv==1.0.1
lxml==5.2.2
